# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tarcza Polska API",
        "description": "Backend for the Tarcza Polska civil resilience platform.\n\n* `/api/v1/*` - Citizen mobile app (Flutter). Authenticate with the JWT returned by `POST /api/v1/devices`.\n* `/api/command/*` - Command Center. Authenticate with the JWT returned by `POST /api/command/login`.\n\nAll spatial payloads are GeoJSON (WGS84, `[lng, lat]` order). Every error is an `ErrorResponse`.\n`GET /map`, `GET /verifications/pending` and `GET /alerts` send an `ETag`; repeat the value in\n`If-None-Match` to get `304 Not Modified`.\n",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/command/incidents": {
            "get": {
                "tags": [
                    "Command"
                ],
                "summary": "Incident feed for the Command Center (open first)",
                "operationId": "get_api_command_incident_list",
                "parameters": [
                    {
                        "name": "all",
                        "in": "query",
                        "description": "1 = include resolved",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident summaries"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient role (error.code: forbidden)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/command/incidents/{id}": {
            "get": {
                "tags": [
                    "Command"
                ],
                "summary": "Full incident detail: raw reports, cells, sources, confidence breakdown (audit-logged)",
                "operationId": "get_api_command_incident_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident detail"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient role (error.code: forbidden)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Open incidents (optionally only those containing my position)",
                "operationId": "get_api_incident_list",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of public incident views",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IncidentView"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/incidents/{id}": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Public view of an incident (aggregated, no raw report positions)",
                "operationId": "get_api_incident_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentView"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/reports": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Submit a report (type -> location -> optional description)",
                "operationId": "post_api_report_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted; clustering happens asynchronously",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportAccepted"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "More than 10 reports in 10 minutes from this device (error.code: too_many_requests, Retry-After header)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/types": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Report categories with labels (for the report screen)",
                "operationId": "get_api_report_types",
                "responses": {
                    "200": {
                        "description": "List of types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReportTypeOption"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Status of one of my reports (which incident it joined)",
                "operationId": "get_api_report_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportStatusView"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorViolation": {
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "description": "Name of the body field (maps 1:1 onto the form field)",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorDetail": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "Stable machine-readable code: bad_request, unauthorized, token_expired, forbidden, not_found, conflict, gone, verification_already_answered, verification_expired, validation_failed, too_many_requests, internal_error.\n",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "retryAfter": {
                        "description": "Seconds until the next attempt may succeed (429 only, also sent as the Retry-After header)",
                        "type": "integer"
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorViolation"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                },
                "type": "object"
            },
            "GeoJsonGeometry": {
                "description": "GeoJSON geometry in WGS84, coordinates ordered [lng, lat].",
                "required": [
                    "type",
                    "coordinates"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "Point",
                            "Polygon",
                            "MultiPolygon"
                        ]
                    },
                    "coordinates": {
                        "description": "Point: [lng, lat]; Polygon: [[[lng, lat], ...]]; MultiPolygon: [[[[lng, lat], ...]]]",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
            "IncidentStatus": {
                "type": "string",
                "enum": [
                    "detected",
                    "verifying",
                    "active",
                    "resolved"
                ]
            },
            "ConfidenceLevel": {
                "type": "string",
                "enum": [
                    "unverified",
                    "likely",
                    "high",
                    "confirmed"
                ]
            },
...
            "Community": {
                "required": [
                    "reports",
                    "answers"
                ],
                "properties": {
                    "reports": {
                        "type": "integer"
                    },
                    "answers": {
                        "type": "integer"
                    },
                    "agreementPct": {
                        "description": "Share of answers confirming the problem; null until somebody answered",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "IncidentSummary": {
                "description": "Public incident fields shared by IncidentView and the map feature properties.",
                "required": [
                    "id",
                    "type",
                    "typeLabel",
                    "status",
                    "statusLabel",
                    "confidenceLevel",
                    "confidenceLabel",
                    "confidenceScore",
                    "startedAt",
                    "lastActivityAt",
                    "community"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ReportType"
                    },
                    "typeLabel": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/IncidentStatus"
                    },
                    "statusLabel": {
                        "type": "string"
                    },
                    "confidenceLevel": {
                        "$ref": "#/components/schemas/ConfidenceLevel"
                    },
                    "confidenceLabel": {
                        "type": "string"
                    },
                    "confidenceScore": {
                        "type": "number",
                        "format": "float",
                        "maximum": 1,
                        "minimum": 0
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastActivityAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastConfirmedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "community": {
                        "$ref": "#/components/schemas/Community"
                    },
                    "summary": {
                        "description": "AI research summary, when available",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "IncidentView": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/IncidentSummary"
                    },
                    {
                        "required": [
                            "area"
                        ],
                        "properties": {
                            "area": {
                                "description": "Polygon / MultiPolygon of verified cells; null while the incident is only detected",
                                "nullable": true,
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/GeoJsonGeometry"
                                    }
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ReportTypeOption": {
                "required": [
                    "value",
                    "label"
                ],
                "properties": {
                    "value": {
                        "$ref": "#/components/schemas/ReportType"
                    },
                    "label": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ReportAccepted": {
                "required": [
                    "reportId",
                    "h3Cell",
                    "createdAt"
                ],
                "properties": {
                    "reportId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "h3Cell": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ReportIncidentRef": {
                "required": [
                    "id",
                    "type",
                    "typeLabel",
                    "status",
                    "statusLabel",
                    "confidenceLevel",
                    "confidenceLabel",
                    "confidenceScore"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ReportType"
                    },
                    "typeLabel": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/IncidentStatus"
                    },
                    "statusLabel": {
                        "type": "string"
                    },
                    "confidenceLevel": {
                        "$ref": "#/components/schemas/ConfidenceLevel"
                    },
                    "confidenceLabel": {
                        "type": "string"
                    },
                    "confidenceScore": {
                        "type": "number",
                        "format": "float",
                        "maximum": 1,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
            "ReportStatusView": {
                "required": [
                    "reportId",
                    "type",
                    "typeLabel",
                    "createdAt",
                    "incident"
                ],
                "properties": {
                    "reportId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ReportType"
                    },
                    "typeLabel": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "incident": {
                        "description": "null until clustering (asynchronous, usually within seconds) attached the report to an incident",
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ReportIncidentRef"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "ReportType": {
                "type": "string",
                "enum": [
                    "power_outage",
                    "water_outage",
                    "fuel_shortage",
                    "road_blocked",
                    "shelter_issue",
                    "other_threat"
                ]
            },
            "CreateReportRequest": {
                "required": [
                    "type",
                    "lat",
                    "lng"
                ],
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/ReportType"
                    },
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "maximum": 90,
                        "minimum": -90
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "maximum": 180,
                        "minimum": -180
                    },
                    "description": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "maxLength": 1000
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}